refactor: renaming for better readability; no re-exports
